refactor form steps and add to zod schema
