+++ src/Components/LetterBuilder/ProgressBar/ProgressBar.scss
