+++ src/Components/LetterBuilder/ProgressBar/ProgressBar.tsx
