--- src/Components/Pages/LetterSender/MultiStepForm.tsx
@@ -1,488 +1,0 @@
-import {
-  Controller,
-  FieldPath,
-  Resolver,
-  SubmitHandler,
-  useForm,
-  UseFormReturn,
-} from "react-hook-form";
-import z from "zod";
-import {
-  Button,
-  FormGroup,
-  RadioButton,
-  TextInput,
-} from "@justfixnyc/component-library";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { useState } from "react";
-import {
-  formatPhoneNumber,
-  handleFormNoDefault,
-  parseFormattedPhoneNumber,
-} from "../../../form-utils";
-import { ProgressBar } from "./ProgressBar";
-import { GeoSearchInput } from "../../GeoSearchInput/GeoSearchInput";
-import { useGetLandlordData } from "../../../api/hooks";
-import { LandlordContact, LandlordData } from "../../../types/APIDataTypes";
-import { Tenants2ApiFetcherVerifyAddress } from "../../../api/helpers";
-
-const FormSchema = z.object({
-  firstName: z.string().min(1, "Name is required for the letter"),
-  lastName: z.string().min(1, "Name is required for the letter"),
-  phone: z
-    .string({
-      error: (iss) =>
-        iss.input === undefined || iss.input === ""
-          ? "Phone number is required for follow up"
-          : "Please enter a complete US phone number",
-    })
-    .length(10, "Please enter a complete US phone number"),
-  // Email has god default validator regex, plus alternatives, and way to
-  // include custom ({pattern: /<regex>/})
-  email: z.email({
-    // To customize error messages add a function that checks different possible
-    // issues with the field, since z.email("invalid format") would also give
-    // the same message even if empty input
-    error: (iss) =>
-      iss.input === undefined || iss.input === ""
-        ? "Email is required to send your copy of the letter"
-        : "Please enter a valid email",
-  }),
-  address: z.object({
-    houseNumber: z.string(),
-    streetName: z.string(),
-    borough: z.string(),
-    zip_code: z
-      .string()
-      .regex(
-        /^\d{5}((-)?\d{4})?$/,
-        "ZIP Code must be of format 12345 or 12345-1234"
-      )
-      .optional(),
-    bbl: z.string().regex(/^\d{10}$/),
-  }),
-  landlordDetails: z.object({
-    primary_line: z.string(),
-    secondary_line: z.string(),
-    city: z.string(),
-    state: z.string(),
-    zip_code: z
-      .string()
-      .regex(
-        /^\d{5}((-)?\d{4})?$/,
-        "ZIP Code must be of format 12345 or 12345-1234"
-      ),
-    urbanization: z.string().optional(),
-  }),
-  // TODO: add fieldArray of issues (see https://react-hook-form.com/docs/usefieldarray)
-});
-
-type FormFields = z.infer<typeof FormSchema>;
-
-interface Step {
-  id: string;
-  name: string;
-  fields?: FieldPath<FormFields>[];
-}
-
-const steps: Step[] = [
-  {
-    id: "Step 1",
-    name: "Contact information",
-    fields: ["email", "phone", "firstName", "lastName"],
-  },
-  {
-    id: "Step 2",
-    name: "Your address",
-    fields: ["address"],
-  },
-  {
-    id: "Step 3",
-    name: "Landlord details",
-    fields: ["landlordDetails"],
-  },
-  { id: "Step 4", name: "Complete" },
-];
-
-export const MultiStepForm: React.FC = () => {
-  const formHookReturn = useForm<FormFields>({
-    // Issue with the inferred type being "unknown" when preprocess() is used to
-    // handle values that should be changed to undefined
-    resolver: zodResolver(FormSchema) as Resolver<FormFields>,
-    mode: "onSubmit",
-  });
-  const { reset, watch, trigger, handleSubmit, setError } = formHookReturn;
-
-  const [currentStep, setCurrentStep] = useState(0);
-
-  const processForm: SubmitHandler<FormFields> = (data: FormFields) => {
-    console.log({ data });
-    reset();
-  };
-
-  const verifyAddress: SubmitHandler<FormFields> = async (data: FormFields) => {
-    try {
-      const verification = await Tenants2ApiFetcherVerifyAddress(
-        "/gceletter/verify_address",
-        { arg: data.landlordDetails }
-      );
-      const isUndeliverable =
-        verification.deliverability !== "deliverable" &&
-        !verification.valid_address;
-      if (isUndeliverable) {
-        setError("landlordDetails", {
-          type: "lob_verification",
-          message: "Landlord address as entered is not deliverable by USPS",
-        });
-      }
-    } catch (error) {
-      // Handle network or other errors
-      console.error("Unable to check address deliverability: ", error);
-    }
-  };
-
-  console.log({ landlordDetails: watch("landlordDetails") });
-
-  const next = async () => {
-    const fields = steps[currentStep].fields;
-
-    // TODO: clean up logic to handle final confirmation step with no fields
-    if (!fields) {
-      setCurrentStep((step) => step + 1);
-      return;
-    }
-
-    // console.log(FormSchema.safeParse(getValues()));
-
-    const output = await trigger(fields, { shouldFocus: true });
-
-    if (!output) return;
-
-    if (fields.includes("landlordDetails")) {
-      // TODO: need to change how steps work when the landlord details step can
-      // have two sub-steps (lookup address then manual enter)
-      await handleSubmit(verifyAddress)();
-      // return;
-    }
-
-    if (currentStep < steps.length - 1) {
-      if (currentStep === steps.length - 2) {
-        await handleSubmit(processForm)();
-      }
-      setCurrentStep((step) => step + 1);
-    }
-  };
-
-  const prev = () => {
-    if (currentStep > 0) {
-      setCurrentStep((step) => step - 1);
-    }
-  };
-
-  return (
-    <form onSubmit={handleFormNoDefault(next)} className="letter-form">
-      <h3>
-        <pre>MultiStepForm</pre>
-      </h3>
-      <p>Multi step state management, customized errors</p>
-      <ProgressBar steps={steps} currentStep={currentStep} />
-      <div className="letter-form__content">
-        {currentStep === 0 && <UserDetailsStep {...formHookReturn} />}
-        {currentStep === 1 && <UserAddressStep {...formHookReturn} />}
-        {currentStep === 2 && <LandlordDetailsStep {...formHookReturn} />}
-      </div>
-      <div className="letter-form__buttons">
-        {currentStep > 0 && (
-          <Button variant="tertiary" labelText="Back" onClick={prev} />
-        )}
-        <Button
-          labelText={currentStep < steps.length - 1 ? "Next" : "Submit"}
-          type="submit"
-        />
-      </div>
-    </form>
-  );
-};
-
-const getOwnerContacts = (
-  data?: LandlordData
-): LandlordContact[] | undefined => {
-  // TODO: review LOC methodology more closely
-  // https://github.com/JustFixNYC/tenants2/blob/master/loc/landlord_lookup.py
-  if (data === undefined) return;
-
-  const owners = data.allcontacts
-    .filter((contact) =>
-      ["HeadOfficer", "IndividualOwner", "JointOwner"].includes(contact.title)
-    )
-    // Sort by title in above order, just happens to be alpha order by default
-    .sort((a, b) => a.title.localeCompare(b.title));
-  return owners;
-};
-
-const formatLandlordAddress = (addr: LandlordContact["address"]): string => {
-  return `${addr.housenumber} ${addr.streetname}${
-    addr.apartment ? " " + addr.apartment : ""
-  }, ${addr.city}, ${addr.state} ${addr.zip}`;
-};
-
-const landlordContactLOBAddress = (
-  addr: LandlordContact["address"]
-): FormFields["landlordDetails"] => {
-  return {
-    primary_line: `${addr.housenumber} ${addr.streetname}`,
-    secondary_line: addr.apartment,
-    city: addr.city,
-    state: addr.state,
-    zip_code: addr.zip,
-  };
-};
-
-const defaultLandlordDetails = {
-  primary_line: "",
-  secondary_line: undefined,
-  city: "",
-  state: "",
-  zip_code: "",
-};
-
-type FormHookProps = UseFormReturn<FormFields>;
-
-const UserAddressStep: React.FC<FormHookProps> = (props) => {
-  const {
-    setValue,
-    setError,
-    formState: { errors },
-  } = props;
-  return (
-    <GeoSearchInput
-      // extra keys (eg. longLat) are ignored
-      onChange={(addr) => setValue("address", addr)}
-      invalid={!!errors.address}
-      setInvalid={(isError) => {
-        if (isError) {
-          setError("address", {
-            type: "custom",
-            message: "Error with address selection",
-          });
-        }
-      }}
-    />
-  );
-};
-
-const LandlordDetailsStep: React.FC<FormHookProps> = (props) => {
-  const {
-    register,
-    control,
-    formState: { errors },
-    getValues,
-  } = props;
-
-  const [lookupComplete, setLookupComplete] = useState(false);
-
-  const {
-    data: landlordData,
-    isLoading,
-    error,
-  } = useGetLandlordData(getValues("address.bbl"));
-
-  const owners = getOwnerContacts(landlordData);
-
-  const showLookup = !isLoading && !error && owners && !lookupComplete;
-  const showManual = !isLoading && (!landlordData || lookupComplete);
-
-  return (
-    <>
-      {isLoading && <>Loading...</>}
-      {error && <>Failed to lookup landlord information</>}
-      {showLookup && (
-        <>
-          <FormGroup
-            legendText="Are any of these your landlord's information?"
-            key="lookup-suggestions"
-          >
-            {errors?.landlordDetails && (
-              <span className="error">{errors?.landlordDetails?.message}</span>
-            )}
-            {owners.map((owner, index) => (
-              // TODO: should update JFCL to allow label to be string or ReactNode
-
-              <Controller
-                name="landlordDetails"
-                control={control}
-                render={({ field }) => (
-                  <RadioButton
-                    {...field}
-                    value={JSON.stringify(
-                      landlordContactLOBAddress(owner.address)
-                    )}
-                    checked={
-                      JSON.stringify(field.value) ===
-                      JSON.stringify(landlordContactLOBAddress(owner.address))
-                    }
-                    onChange={() =>
-                      field.onChange(landlordContactLOBAddress(owner.address))
-                    }
-                    labelText={`${owner.value}: ${formatLandlordAddress(
-                      owner.address
-                    )}`}
-                    id={`landlord-address_${index}`}
-                    key={index}
-                  />
-                )}
-              />
-            ))}
-            <Controller
-              name="landlordDetails"
-              control={control}
-              render={({ field }) => (
-                <RadioButton
-                  {...field}
-                  value={JSON.stringify(defaultLandlordDetails)}
-                  checked={
-                    JSON.stringify(field.value) ===
-                    JSON.stringify(defaultLandlordDetails)
-                  }
-                  onChange={() => field.onChange(defaultLandlordDetails)}
-                  labelText="None of the above"
-                  id={`landlord-address_none`}
-                  key="none-option"
-                />
-              )}
-            />
-          </FormGroup>
-          <Button
-            labelText="Manually enter"
-            type="button"
-            onClick={() => setLookupComplete(true)}
-          />
-        </>
-      )}
-      {showManual && (
-        <FormGroup
-          legendText="Please provide your landlord's mailing address"
-          invalid={!!errors?.landlordDetails}
-          invalidText={errors?.landlordDetails?.message}
-          key="manual-input"
-        >
-          <TextInput
-            {...register("landlordDetails.primary_line")}
-            id={`form-landlord-primary-line`}
-            labelText="Primary line"
-            invalid={!!errors?.landlordDetails?.primary_line}
-            invalidText={errors?.landlordDetails?.primary_line?.message}
-            invalidRole="status"
-            type="text"
-          />
-          <TextInput
-            {...register("landlordDetails.secondary_line")}
-            id={`form-landlord-secondary-line`}
-            labelText="Secondary line (optional)"
-            invalid={!!errors?.landlordDetails?.secondary_line}
-            invalidText={errors?.landlordDetails?.secondary_line?.message}
-            invalidRole="status"
-            type="text"
-          />
-          <TextInput
-            {...register("landlordDetails.city")}
-            id={`form-landlord-city`}
-            labelText="City/Borough"
-            invalid={!!errors?.landlordDetails?.city}
-            invalidText={errors?.landlordDetails?.city?.message}
-            invalidRole="status"
-            type="text"
-          />
-          {/* TODO: use dropdown for state to ensure correct format */}
-          <TextInput
-            {...register("landlordDetails.state")}
-            id={`form-landlord-state`}
-            labelText="State"
-            invalid={!!errors?.landlordDetails?.state}
-            invalidText={errors?.landlordDetails?.state?.message}
-            invalidRole="status"
-            type="text"
-          />
-          {getValues("landlordDetails.state") === "PR" && (
-            <TextInput
-              {...register("landlordDetails.urbanization")}
-              id={`form-landlord-urbanization`}
-              labelText="Urbanization (Puerto Rico only)"
-              invalid={!!errors?.landlordDetails?.urbanization}
-              invalidText={errors?.landlordDetails?.urbanization?.message}
-              invalidRole="status"
-              type="text"
-            />
-          )}
-          <TextInput
-            {...register("landlordDetails.zip_code")}
-            id={`form-landlord-zip-code`}
-            labelText="ZIP Code"
-            invalid={!!errors?.landlordDetails?.zip_code}
-            invalidText={errors?.landlordDetails?.zip_code?.message}
-            invalidRole="status"
-            type="text"
-          />
-        </FormGroup>
-      )}
-    </>
-  );
-};
-
-const UserDetailsStep: React.FC<FormHookProps> = (props) => {
-  const {
-    register,
-    control,
-    formState: { errors },
-  } = props;
-  return (
-    <>
-      <TextInput
-        {...register("firstName")}
-        id="form-firstName"
-        labelText="First name"
-        invalid={!!errors?.firstName}
-        invalidText={errors?.firstName?.message}
-        invalidRole="status"
-        type="text"
-      />
-      <TextInput
-        {...register("lastName")}
-        id="form-lastName"
-        labelText="Last name"
-        invalid={!!errors?.lastName}
-        invalidText={errors?.lastName?.message}
-        invalidRole="status"
-        type="text"
-      />
-      <TextInput
-        {...register("email")}
-        id="form-email"
-        labelText="Email"
-        invalid={!!errors?.email}
-        invalidText={errors?.email?.message}
-        invalidRole="status"
-        type="email"
-      />
-      <Controller
-        name="phone"
-        control={control}
-        render={({ field }) => (
-          <TextInput
-            {...field}
-            value={formatPhoneNumber(field.value)}
-            onChange={(e) =>
-              field.onChange(parseFormattedPhoneNumber(e.target.value))
-            }
-            id="form-phone"
-            labelText="Phone"
-            invalid={!!errors?.phone}
-            invalidText={errors?.phone?.message}
-            invalidRole="status"
-            type="tel"
-          />
-        )}
-      />
-    </>
-  );
-};
